@@ -5,31 +5,42 @@ import Center from "../../components/Center"
 import Divider from "../../components/Divider"
 import HeaderTop from "../../components/HeaderTop"
 import { loading } from "../../Global"
+import { user } from "../../network/Endpoint"
+import { useNavigate } from "react-router-dom";
+
 import './admin.scss'
+
 
 export default class Admin extends React.Component {
     state = {
         fillPercent: 0,
         show: false,
+        email: "",
+        password: "",
     }
 
-    login() {
+    async login() {
         if (loading) {
-            var promise = new Promise(resolve => {
-                setTimeout(resolve, 10000)
-            })
-            loading.current?.showWithPromise(promise);
+            const p = user.login(this.state.email, this.state.password)
+            const successful = await loading.current?.showWithPromise(p)
+            const navigate = useNavigate();
+                navigate("/admin")
+            if (successful) {
+                console.log("log in success")
+                const navigate = useNavigate();
+                navigate("/admin")
+            }
         }
     }
 
-    renderInput(label: string, inputType: string, placeholder?: string) {
+    renderInput(label: string, inputType: string, value: string, onChange: (s: string) => void, placeholder?: string) {
         if (!placeholder) {
             placeholder = `请输入${label}`
         }
         return (
             <div className="label-input">
                 <span className="label">{label}</span>
-                <input type={inputType} placeholder={placeholder} style={{ flex: 7 }} />
+                <input type={inputType} placeholder={placeholder} style={{ flex: 7 }} value={value} onChange={e => onChange(e.target.value)}/>
             </div>
         )
     }
@@ -47,8 +58,8 @@ export default class Admin extends React.Component {
                     </div>
 
                     <div className="login-box">
-                        {this.renderInput("用户名", "email")}
-                        {this.renderInput("密码", "password")}
+                        {this.renderInput("用户名", "email", this.state.email, email => this.setState({email}))}
+                        {this.renderInput("密码", "password", this.state.password, password => this.setState({password}))}
                         <button
                             onClick={this.login.bind(this)}
                             className="login-btn"
